Revert "feat: 게임을 준비하는 상태의 기능 구현"
This reverts commit f1de685719152e7ba169a5ccbcf49f46cfef70bc.
