+++ src/test/java/chess/controller/command/StartCommandTest.java
@@ -5,7 +5,6 @@ import chess.domain.board.BoardFactory;
 import chess.domain.piece.Team;
 import chess.repository.BoardDao;
 import chess.repository.InMemoryBoardDao;
-import chess.view.OutputView;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
@@ -41,8 +40,22 @@ class StartCommandTest {
     }
 
     @ParameterizedTest
-    @ValueSource(strings = {"start", "emd", "sta rt"})
-    void move나_status를_입력받지_않으면_예외가_발생한다(String command) {
+    @ValueSource(strings = {"start", "start ", " start", "Start", " start  "})
+    void start를_입력받으면_MoveCommand_객체가_반환된다(String command) {
+        Command startCommand = new StartCommand(boardDao);
+        List<String> input = Arrays.stream(command.split(" "))
+                .map(String::trim)
+                .filter(x -> !x.isEmpty())
+                .collect(Collectors.toList());
+
+        Command result = startCommand.execute(input);
+
+        assertThat(result.isSameType(CommandType.MOVE)).isTrue();
+    }
+
+    @ParameterizedTest
+    @ValueSource(strings = {"start1", "star", "sta rt"})
+    void start를_입력받지_않으면_예외가_발생한다(String command) {
         Command startCommand = new StartCommand(boardDao);
         List<String> input = Arrays.stream(command.split(" "))
                 .map(String::trim)
@@ -54,21 +67,8 @@ class StartCommandTest {
     }
 
     @ParameterizedTest
-    @ValueSource(strings = {"status  a2"})
-    void status명령어는_명령어만_입력_가능하다(String command) {
-        Command startCommand = new StartCommand(boardDao);
-        List<String> input = Arrays.stream(command.split(" "))
-                .map(String::trim)
-                .collect(Collectors.toList());
-
-        assertThatThrownBy(() -> startCommand.execute(input))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("status 명령어는 값을 하나만 입력해야합니다.");
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = {"move a2"})
-    void move명령어는_도착지와_출발지의_정보를_입력해야한다(String command) {
+    @ValueSource(strings = {"start  2"})
+    void start명령어는_명령어만_입력_가능하다(String command) {
         Command startCommand = new StartCommand(boardDao);
         List<String> input = Arrays.stream(command.split(" "))
                 .map(String::trim)
